GafferScene : Don't link against IECoreAlembic
Since the AlembicInput node was removed, GafferScene no longer needs any symbols from IECoreAlembic. We now use a startup file to load IECoreAlembic in the same was as we do for IECoreUSD.

For some reason this change was necessary to allow `SceneReaderTest.testAlembic` to pass on Travis, although it was passing OK for me locally. Perhaps some issue whereby the linker on Travis was optimising away unused libraries?