AuxiliaryConnectionsGadget : Fix lifetime issues
These were revealed by a crash running the unit tests on Mac. The issue was that the python bindings were hanging on to an AuxiliaryConnectionGadget after its parent GraphGadget had died. This meant that m_graphGadget was a dangling pointer. We now avoid the temptation of storing `m_graphGadget` at all, and use `parentChanging()` to detach from all signals so the AuxiliaryConnectionsGadget is inactive when it has no parent.