PathListingWidget : Optimise `setExpansion()`
The key here was to reverse the logic. Searching for the index for a particular path is incredibly slow, because we must do a linear search of the children of the PathModel::Item at each level of the path. On the other hand, doing PathMatcher lookups is very fast, because that is what PathMatchers are optimised for. So we now recurse through the model hierarchy directly, doing tests against the PathMatcher as we go, pruning the recursion when we detect that there are no descendant matches.

This gives a greater than 100x speedup when transferring the expansion of around 35000 instances into the SceneHierarchy.