PathListingWidget : Optimise `getSelection()`
It turns out that `QItemSelectionMode::selectedRows()` is insanely slow for a large selection, even when your selection is guaranteed only to contain complete rows anyway. This simple change gives us a 180x speedup when querying a selection of ~35000 items.

The pattern emerging here and in previous commits seems to be that if you want to avoid performance traps in Qt, you need to avoid uttering "QItemSelectionModel" and "rows" in the same sentence.