Wrapper classes : Translate all Python exceptions
Wrapper classes implement the transfer of control from C++ into Python when a Python class overrides a C++ virtual method. If a Python exception is thrown when we invoke the method override, this is automatically converted into a `boost::python::error_already_set` exception for us, with the Python error indicator set with the full information from this exception. If we don't catch this exception in the wrapper, two bad things happen :

- The C++ calling code is faced with an `error_already_set` exception that it doesn't know how to handle.
- The Python error indicator remains set, which can cause very strange errors when we next enter Python, with the earlier exception appearing to have been thrown from the later code.

By translating the Python exception to a pure C++ exception we avoid the first problem while also clearing the Python error indicator, thus avoiding the second problem too.