PlaybackTest : Ensure that current frame is serialised to script
The Playback class is what the timeline uses to drive the context. It drives contexts rather than plugs on the ScriptNode because a) This allows it to be used with custom contexts and b) Code should not modify `ScriptNode::framePlug()`. Here we just check that the changes to the context are accounted for when serialising the script.