Serialisation : Improve support for Python-based serialisers
- Allow Python method overrides to call the base class implementation
- Allow subclassing of NodeSerialiser

Because neither of these things were possible before, some downstream projects have had to resort to implementing serialisers which duplicate the logic of the NodeSerialiser. This adds unnecessary complexity
and makes them prone to breaking when the details of Gaffer's own serialisers change.

The key change here is documented at some length in SerialisationBinding.inl. This is the first time we've
successfully allowed subclassing in Python while still allowing some C++ subclasses to remain completely
private. Hopefully we can now use this same technique to improve our wrapping of classes such as ComputeNode
and Gadget.