ExpressionUI : Fix initial UI state
When an Expression node is made directly via the Node menu, it starts out without either the expression or language set. If you entered an expression _before_ choosing the language, then this would do nothing, and then when choosing the language you'd lose the expression! I presume this hasn't been spotted until now because most people create expressions via the NodeEditor right click menu which sets the language upfront.

The fix applied here is simple : if the language isn't set, label the menu to prompt the user to set it, and don't enabled the expression text field until that is done.