Viewer : Add `viewContextMenuSignal()`
This allows context menus for views to be customised using the same basic mechanism we use in the NodeGraph and elsewhere.

The low level mechanism for creating context menus in Widget now ignores events with modifiers, so that _Alt + Right Drag_ can still do camera movement rather than opening the context menu.