FormatPlug: fix hard crashes if None is passed through python binding
Problem can be replicated by executing

import GafferImage
GafferImage.FormatPlug.acquireDefaultFormatPlug( None )

in the ScriptEditor.

While that seems like a stupid thing to do, I had an issue with naively calling
acquireDefaultFormatPlug in a node's initializer, passing in self.scriptNode()
which obviously is None during node creation. Easy enough to fix, but the hard
crash is avoidable and the message gives a hint what went wrong.