ImageAlgo : Simplify `parallelGatherTiles()` using lambdas
This lets us keep the code local to where it is called, and omit a bunch of boilerplate with constructors and members which are now replaced by simple captures. It also allows us to automatically infer the result type for the TileFunctor using `std::result_of`, so that callers no longer need to provide a `TileFunctor::Result` typedef.