ImageAlgo : Reimplement `parallelProcessTiles()`
We now use the same mechanism used for `parallelGatherTiles()`, just without the final serial step. As a side effect of this, the tiles are now processed in a top to bottom order by default, which tends to work better when processing scanline image files loaded by the ImageReader. See #2380 for discusssion. The tile order can be controlled by the new `tileOrder` parameter.