ImageAlgo : Make parallel methods lambda friendly
This means accepting lambdas defined as temporaries within the function call itself. The original `Functor &` arguments can't accept temporaries, so we must change the signature. For `parallelGatherTiles()` we use `const TileFunctor &` to accept temporaries and also enforce that the functor doesn't have mutable internal state, since modifying state is best done in the serial gather step.

For the other arguments we use `Functor &&`, which is a "universal reference" which magically turns into an rvalue reference or lvalue reference, depending on what is passed. This seems to be the only way of accepting a non-const temporary without passing by value (requiring the argument to be copyable). This seems to be the mind-bending underpinning of C++11's surface loveliness, and I found this reference to be useful in coming to some sort of understanding :

https://isocpp.org/blog/2012/11/universal-references-in-c11-scott-meyers

Breaking Change :

- Image Algo : Changed signatures for `parallelProcessTiles()` and `parallelGatherTiles()`