GafferOSL::ShadingEngine : Simplify implementation using flat_map
This is equivalent to the manually sorted vectors we were using before, but easier to use. Also renamed `findDebugResult()` to `acquireDebugResult()` to more accurately reflect its role. The usual semantics of `find()` are that it may fail, and that is not the case here. In Gaffer we use "acquire" everywhere we want to find something and create it if it does not yet exist.