OSL shaders : Add OutUV shader
This allows UV sets to be created. Although OSL does now provide a vector2 type via a contributed header, I've used a regular 3d vector for the input type, because I'm yet to be convinced that it's worth our while to provide 2 and 3 dimensional variants of every vector related shader. I also haven't added OutPoint2, OutVector2 etc, because I'm not sure what the use case would be, and we'll shortly be moving away from the Out* shaders in favour of @danieldresser's OSLCode style overhaul of OSLObject.