ShadingEngine : Support reading time and context variables
Time is available via the `time` global variable, and receives its value from `Context->getTime()` in the same way as it does in OSLExpressionEngine. Note : This is likely not the same as in a renderer, where time is likely to be frame-based. Is this what we want?

Context variables are available via a `getattribute()` call with a scope of "gaffer:context", in the same way as they are in OSLExpressionEngine.