OSL ShadingEngine : Fix bug caused by TBB cancellation propagation
This manifested itself as tiles with unshaded regions being created very sporadically during the cancellation of image generation in the viewer. Despite trying, I've been unable to demonstrate this with an automated test case, so all I have to offer is the following explanation, and the fact that after making the fix, I have yet to see a bad tile.

The issue was with TBB's cancellation mechanism and its interaction with nested parallelism, as documented here :

https://www.threadingbuildingblocks.org/docs/help/tbb_userguide/Cancellation_and_Nested_Parallelism.html

In the case in hand, the ImageGadget is using `parallelGatherTiles` to compute tiles in parallel, and nested within that, ShadingEngine is using `parallel_for` to compute pixels in parallel within individual tiles. This can be represented with the following simplified task tree :

            Root
             / \
            /   \
           /     \
       Tile A   Tile B          (ImageGadget tasks)
        /  \      / \
       /   ...  ...  \
      /               \
    Batch A1         Batch B1   (ShadingEngine tasks)

By default, TBB treats this as a single task tree for the purposes of cancellation. In TBB terms, cancellation means that if an exception is thrown by _any_ task in the tree, then no new tasks in the tree will be started. If Batch B1 throws `IECore::Cancelled`, then Batch A1 may not even be run, even though it's `parallel_for` has been called and the task is ready to go. Worse, the `parallel_for` for Batch A1 returns silently, as if all is well, so we return an incomplete tile to the caller without even being aware that something has gone wrong.

The more general issue is that TBB's defaults don't work well when parallelism is hidden behind library interfaces. It is merely an implementation detail of `ShadingEngine::shade()` that it uses `parallel_for` internally, and ShadingEngine has no knowledge of how it may be embedded in an outer task tree. The solution is to insulate `ShadingEngine` from any outer nesting level by preventing TBB cancellation from propagating down to its tasks.

In practice, once the fix is made, Batch A1 in our example will throw `IECore::Cancelled` as soon as it is run anyway. But the key difference is that this means `ShadingEngine::shade()` fails for Tile A and we don't return the tile to the caller.

The upshot of all this is that we need to apply the same fix _everywhere_ we've hidden TBB parallelism behind a library interface in Gaffer or Cortex.