BackgroundTask : Fix cancellation of yet-to-be-started tasks
In specific circumstances the TBB scheduler gives total priority to already-running recursive tasks, refusing to run our enqueued task until it has exhausted the supply of spawned recursive tasks. This could mean the UI thread was waiting for cancellation of a single enqueued color inspector task, but that task was being constantly passed over in favour of ImageGadget tile tasks. The end result : the UI stalls until all tiles have been generated.

The fix is to allow return from `BackgroundTask::cancelAndWait()` _before_ the enqueued TBB task has been started. We use a new status field to mark the task as cancelled so that it becomes a no-op when it does eventually run.