RankFilter : Support cancellation in inner loop
I had hoped we might be able to perform the check only once for each scanline within a tile, but once per pixel is necessary when radius is in the 1000s, because that means each output pixel processes millions of input pixels. The checks add a 2-4% overhead over the original. In future we can probably remove this overhead for batch rendering by dispatching to a specialised implementation when there is no `IECore::Canceller` present. But I'm not inclined to add that complexity for these few percent; in future we should optimise the RankFilter anyway, at which point the overhead would become relatively more significant, and we could do the specialised dispatch as part of the optimisation work.