Resample : Add cancellation checks
For the separable cases, it is sufficient to check just once per scanline within a tile, but for the non-separable case we must check once per pixel because each pixel can require a significant amount of work. Testing with a 200 pixel Blur (which uses a separable Resample internally) this shows negligible overhead.