BusyWidget : Add `setBusy()` and `getBusy()` methods
To date we have controlled BusyWidgets by managing their visibility, but this causes widget layouts to be recomputed, which might not be desirable.

This commit also fixes the widget at the size requested in the constructor, rather than letting it expand. Expansion doesn't make any sense because the graphic doesn't expand to use the space, and the same effect can be achieved using the parent container.