BackgroundMethod : Fix bug which could mean postCalls were missed
This was causing the following sporadic errors on Travis :

```
testCancelWhenHidden (GafferUITest.BackgroundMethodTest.BackgroundMethodTest) ...
No output has been received in the last 10m0s, this potentially indicates a stalled build or something wrong with the build itself.
```

The issue was that we relied on the background task to schedule the `postCall`, but if cancellation was quick enough it would stop the background task from even starting. Since we've already committed ourselves by calling the `preCall`, we must make sure there is a matching `postCall`.