DispatchDialogue : No longer requiring an ApplicationRoot for the menu
The dispatch dialogue now stores the menu definition directly on the
class, thus avoiding the need to store it as a attached property to
the application.

That has the advantage of being simpler, as well as not requiring
that scripts used by the dialogue be parented to the application.

The dispatch gui.py startup script was updated to reflect the changes.