gui app : Remove BrowserEditor from the available panes.
It hasn't been very useful in its current form. Files and FileSequences modes are more easily used from plug widgets directly on the nodes that will load the files. Nodes mode seems the most promissing for use in the gui app, but it needs significant feature enhancement before it'd be truely useful. Even then, it seems to make sense as a standalone "Graph Browser" pane separate from the other modes.