TweakPlugValueWidget : Fix popup menu
This was calling `isinstance( plug, TweakPlug )`, which spewed errors like so, breaking all plug popup menus :

```
NameError: global name 'TweakPlug' is not defined
```

It was also checking the type of the wrong plug - typically it's the parent plug which is a TweakPlug, because the popup menu is being raised for the child plugs. It should work either way now though.

Also removed a bunch of unused imports.