ArnoldLight : Fix performance problem
The call to `ancestor<Shader>()` would return `nullptr` for most parameters, but before doing so, it had to search all ancestors. And what's the final ancestor? An ApplicationRoot, which happens to be subclassed in Python, turning an innocent `isInstanceOf()` call into an expensive acquisition of the GIL.

The fix is simple. We shouldn't have been searching all ancestors anyway. Instead we should just search for the plug's node (guaranteed to terminate before getting to the ApplicationRoot) and then test to see if it's a `Shader`.

This gives a greater than 5x speedup in scene generation for a synthetic scene consisting of lots of Arnold lights. In the production scene that prompted this investigation, the GIL locks were the third most expensive thing in the profile, so I expect a decent win there too.

(cherry picked from commit 8d2285dde7ed9d7881db1f76318d92c14655b4d4)