CollectScenes : Replace bogus ArrayPlug input with ScenePlug
We appear to have got away with this egregious mistake due to a compatibility config allowing us to perform the expected `collectScenes["in"].setInput( otherNode["out"] )`. The fix is a single line in CollectScenes.cpp, but requires another compatibility config to allow us to load old scripts, which have serialisations referring to children of the old ArrayPlug.