IECoreGLPreviewTest : Use exr for temp image files.
Tiff should be fine, but at IE, we seem to have some sort of config that assumes tiffs are in a weird color space and screws up the test.  We can get around this by using an isolated env, but since most of our test files use exrs anyway, it seems more reliable and consistent to always use exr.