ScriptNodeBinding : Fix GIL management bugs
Now we have asynchronous compute, it is critical that any graph editing operation releases the GIL. This is because the edit will cancel all background tasks and wait for them before proceeding. If the background task is implemented in Python, it will need to acquire the GIL before it can be cancelled, therefore the edit must release the GIL first.

These bugs first manifested in the wild as hangs when cutting and pasting nodes.