ParallelAlgo : Refactor `callOnUIThread()` internals
This has been working perfectly well in the real world, but continuing to cause problems when running the Catalogue unit tests on Travis.

The issue was related to the lack of thread safety in `CallOnUIThreadSignal` and the `ParallelAlgoTest.ExpectedUIThreadCall` we use to emulate the UI for testing purposes. The `ExpectedUIThreadCall` class was sometimes modifying its connection while the signal was still being emitted, with the most common symptom being that a slot was called twice (once by the old connection, once by the new one), or the `CallOnUIThreadSignal` block exited before all expected work was done. This could manifest as either hangs or test failures.

Since using a signal didn't make that much sense anyway (we only want one thing connected at any time) I've replaced it with a simple `registerUIThreadCallHandler()` function, which allows us to manage an internal mutex around the handler.

> Tip : For a long time these problems never manifested locally, only on Travis.
> Using `taskset 0x01 gaffer test ...` made them appear locally too, as it forces
> Gaffer onto a single CPU, which I presume yields threading patterns closer to
> those seen on Travis.