MatchPatternPathFilter : Make exception-safe
`std::remove_if()` has undefined behaviour if an exception is thrown, and in practice we end up with a child vector containing null pointers. This was causing the file browser to crash if pointed at files with strange permissions.

The fix is to catch the exception and turn it into an error. The alternatives would be to allow the exception to be thrown but to use something more robust than `remove_if()`. This doesn't seem at all useful to client code : it wants to be able to show what it can, rather than allow a single bad file to prevent use listing all the others.

Fixes #2798.