Config : Remove InputGenerator compatibility
This dates back to version 0.16.0.0 (August 2015) when we moved a bunch of nodes from using this nasty old InputGenerator thing to using the ArrayPlug that we use today. Removing the compatibility means that files from version 0.15.0.0 will no longer load - they will need to be resaved from a version after 0.16.0.0 and before this commit is merged.

This is a prerequisite for embarking on #2123. Strictly speaking I could probably do some fiddling to keep this compatibility in place, but given the age of the version in question I don't think it's worth the effort.

Breaking Changes :

- Removed compatibility for loading files from Gaffer 0.15.0.0 and earlier. Resave the file
  in a more recent version before loading in the current version. Note that this can also expose
  bugs in custom scripts : if a script incorrectly attempts to connect to an ArrayPlug instead
  this will now fail instead of being silently corrected.