Python files : Remove reliance on switch compatibility configs
There is some subtlety here for the ShaderSwitch case. When ShaderSwitches were first introduced, they always had "in" and "out" plugs added during construction, but in 0.31.0.0 this ceased to be the case, so that we could support multiple parameter types (previously we only supported RSL coshaders). When we did that, we added `shaderSwitchCompatibility.py` to auto-add the old "in"/"out" plugs on first attempted access, and we still have that same compatibility code in our new `switchCompatibility.py` file. The difference is that the new ShaderSwitch stub isn't a true C++ class as before, so the compatibility `__getitem__` is lost through slicing when retrieving the node from C++. The bottom line is this :

```
s = GafferScene.ShaderSwitch()
s["in"] # Works

script = Gaffer.ScriptNode()
script["s"] = GafferScene.ShaderSwitch()
script["s"]["in"] # Fails
```

The compatibility we have now is sufficient for loading old files, but will fail for any custom code which has the latter form above. The existence of such code seems unlikely since to support current shader parameters (rather than RSL coshaders) a call to `setup()` is required. But I felt it was important to document this in case...