GafferScene : Remove FilterSwitch
Note the special handling required in `FilterPlug::sceneAffectsMatch()` to keep the query working when a FilterPlug receives an input from a context-varying Switch. This is similar to the custom handling required in Shader/ShaderPlug : the generic Switch node works perfectly as long as you stay within the confines of the ComputeNode framework, but if you step outside of it you have to accept some additional responsibility. The upshot of this is that it's now essential to call `FilterPlug::sceneAffectsMatch()` rather than `Filter::sceneAffectsMatch()`, so `Filter::sceneAffectsMatch()` is now private.

Breaking Changes :

- GafferScene : Removed FilterSwitch node : use a regular Switch instead. Backwards compatibility
  for old files is provided by converting FilterSwitches during loading.
- Filter : Made `sceneAffectsMatch()` protected. Use the new `FilterPlug::sceneAffectsMatch()` method
  instead.