ShaderBall : Mark child nodes as read only
Otherwise the TransformTool will edit the internal transforms and everything will appear to have worked, but the results will not be saved. Really, I wonder if `childNodesReadOnly` should actually be `true` for all nodes by default, and `false` only for Boxes and ScriptNodes. But for now we'll just make a more minimal change.