OpenGLRenderer : Fix GL resource management threading bug
We need to destroy GL resources on the main GL thread, but the lifetime of the OpenGLAttributes instances is managed by clients of the renderer, and they can destroy them at any time, on any thread. So we use the same method we use to control destruction of OpenGLObjects : maintain a list of everything we've created, and destroy it on the main thread when we are the last owner.

I was a little concerned at the overhead the destruction checks would have on interactive rendering, but profiling shows them to have negligible impact.

Fixes #2809, although since I couldn't reproduce an actual crash, all I've been able to validate is that this does prevent us from destroying OpenGLAttributes instances on the wrong thread.