SceneElementProcessor : Fix over-zealous dirty propagation
If an output has an input connection (pass-through for instance), it can't have a dependency on anything, because there will be no compute for it on this node. This fixes a problem whereby changing a filter could trigger more updates than necessary - for instance, IPR would rehash objects and transforms instead of just attributes.