Viewer : Perform toolbar update in `plugDirtiedSignal()`
There are two things we want to do when the "active" plug on a tool changes value : turn off other tools, and switch to the right toolbar for the new tool. We were originally doing all this in `plugSetSignal()`, but that's not quite right. As the Node documentation states, the graph may be rewired in `plugSetSignal()` (meaning this is the right place to modify other tools), but observers (the toolbars) should perform their updates in `plugDirtiedSignal()`.

This fixes a flickering issue with the `TransformToolUI._SelectionWidget`, where we got an additional update with an empty selection, due to `TransformTool::plugDirtiedSignal()` not having been called yet.