VectorDataWidget : Fix treatment of `header` parameter
This was intended to provide a header value per-data rather than per-column, but was documented poorly. The only usage to date in Gaffer is in `GafferCortexUI.CompoundVectorParameterValueWidget`, which clearly provides only a value per-data. Presumably we have never used this with compound data types because it was broken, but we now ensure that there is a unique header per column by appending ".X", ".Y" etc to the custom header provided.

Also added `set/getHeader()` accessors.