GafferScene config : Make FilterProcessor compatibility work with Boxes
The previous fix only overrode `GafferScene.FilterSwitch.__getitem__`, but when a node is retrieved from a Box, it emerges as just a `Gaffer.Switch`. This is because the Python compatibility shim for `FilterSwitch` isn't a full wrapped class, so the Python part of the type is forgotten on a trip into C++ and back out again. We now override `Gaffer.Switch.__getitem__` and check to see if the first element of the array is a `FilterPlug`, performing the convert-on-load behaviour if it is.