BoxIO : Remove bogus `setup()` call from serialisation
This was a way of triggering the creation of the promoted plug when a BoxIO node was pasted into a Box. It was an abuse of the API because it didn't pass a plug to `setup()` - instead it passed nothing as a secret way of saying "just do the promotion bit". This resulted in serialisations which were misleading to anyone using them for scripting. We now use signals and slots to automatically create the promoted plug when necessary, and we disallow null plugs in calls to `setup()`. A compatibility config turns the old bogus calls into a no-op.

Note that this just fixes the most egregious problem with the serialisation. BoxIO nodes still serialise awkwardly with explicit calls to add the plugs individually, and even to make the internal connection between them.

> Caution : New serialisations made without this `setup()` call will not create the promoted plug if pasted
> into a Box in an older version of Gaffer.