BoxIO : Fix creation of pass-through after loading old file
There's some important subtlety here. We can't just create the Box.enabled plug and connect to it during the loading of the old script, because we don't have a passThrough connection defined for our BoxOut yet. By connecting, we'd be assuming responsibility for defining the pass-through, and we'd have broken any manual pass-throughs that existed previously. The solution is to only promote the enabled plug when we first get a connection to the passThrough plug. This is arguably better even for newly created BoxOuts, as it means we don't give the appearance of disablability to Boxes until the user has actually defined something meaningful.