IE Config : Always prepend DCC includes
This ensures we get the correct includes for the DCC. This was particularly necessary for Houdini 17 and OpenVDB, which includes a custom _sesi namespace.

I'm not sure why I took the cautious approach mentioned in the todo... presumably it was more todo with production deliveries than anything else.

Also setting VDB_LIB_SUFFIX for Houdini builds.