GUI config : Don't register persistent layouts
Persistent layouts get saved back into the user's own preferences, which is definitely not what we want for the standard layouts. This was the cause of problems like #2882, because the new standard layout was leaked into the user's preferences, which then errored when being loaded in an old Gaffer.

A new unit test ensures that we can't make this mistake again, since it's easily done.