LoopUI : Add context menu item for connecting previous to next
Otherwise it's a bit fiddly to make the dots manually and get the nodules on the right edges. The resulting layout is a little, shall we say, jaunty, but I'd like to leave that as a future improvement for StandardGraphLayout. We could force the node positions to what we think they should be, but then we're missing the collision detection that StandardGraphLayout performance, and we're vulnerable to any changes in size/location of NodeGadgets and nodules.