dispatch app : Change parameter semantics
Add -tasks for specifying the task nodes to dispatch.
Change -nodes to specify which node UIs should be shown in -gui mode.

It seems reasonable to use the DispatchDialogue with no visible task nodes, in the case of a pre-made script. This allows the user to differentiate between tasks to dispatch and nodes to display.