"Box Node" doc : Add images, demos, and scripts
Because we don't have a way of automatically creating .gifs or illustrations,
screengrab.py contains an index of the scripts used to create them, with
a list of which image features will be needed before it can be fully
automated. Note that this is a lofty goal: it's unlikely wel'l be able
to automate every kind of illustration or animation.