ViewportGadget/GadgetWidget : Fix enter/leave event handling bug
We now correctly send enter/leave events to child gadgets when the mouse enters and leaves the viewport gadget. Previously we only did so when the mouse moved. This fixes a highlighting bug when showing menus on PlugAdders. When the mouse enters the menu, the PlugAdder now gets a leave event and correctly drops its highlighting.