LocalDispatcher : Remove extraneous arguments to `Job.__init__()`
Everything else is derived from the dispatcher and context inside the constructor, so I don't see any reason to split the rest of the logic out into the caller. The general idea is that the Job encapsulates everything about a single dispatch. Note that I haven't marked this as a breaking change, because the `Job` constructor is only for the use of the dispatcher itself.