AuxiliaryConnectionsGadget : Draw connections to parent nodules
This is particularly useful in conjunction with CompoundNumericNodule. Now, when a color or vector nodule is collapsed so its child connections are not shown fully, the auxiliary connections for them will be drawn to the parent nodule instead of the node. This makes node graphs much easier to read.

The implementation for this is a little tricky, because now we must regenerate all output connections when nodules are added to and removed from the source node. Prior to this, we only needed to consider events on the destination node. To help with this, we use a `multi_index_container` to store connections such that we can get quick access to  connections associated with any given endpoint, be it source or destination, NodeGadget or Nodule.

This new data structure also allows us to fix a bug whereby filtering a node out of the GraphGadget would not correctly remove any output connections.