ImageView : Give internal node a unique TypeId
The comment was bogus. Node TypeIds are used in generating the hash, so it is not acceptable to omit one. I think it's unlikely this is the cause of the color inspector update bug we've been hunting (the symtoms would likely be more severe), but it's another thing to rule out.