TransformTool : Fix crash caused by selection deduplication
When multiple paths are selected, it's possible that they are all affected by the same upstream transform plug, in which case we deduplicate the selection so that each transform plug is included only once. There was a bug in the way we were doing this though; the selection for `lastSelectedPath` could be skipped by the deduplication process, meaning `Selection lastSelection` was never initialised. This lead to us pushing an empty item on to the end of `m_selection`, which then caused null dereferences in `updateHandles()` :

```
Stack trace:
Gaffer::TransformPlug::pivotPlug()
GafferSceneUI::TransformTool::Selection::orientedTransform(GafferSceneUI::TransformTool::Orientation) const
GafferSceneUI::RotateTool::updateHandles(float)
GafferSceneUI::TransformTool::preRender()
```

We now deduplicate more carefully to avoid the problem.

> Note : The conditional/message for the case where we don't find `lastSelectedPath` is
> not a part of the fix. It just addresses a hypothetical situation that I originally thought
> may have been the cause, but was not.