Dot/ContextProcess/Switch/Loop : Serialise using `setup()` method
This makes the serialisations slightly shorter, but more importantly means they are a better example for folks using them to learn the Gaffer scripting API.

I've deliberately allowed some duplication of logic in the serialisers, because the logic is subtly different for Switch anyway, and I'd rather duplicate some small bits of private code than commit to a public API prematurely. Added todos to this effect, with some thoughts as to an alternative approach.