ShaderAssignment : Assign "osl:shader" as "osl:surface"
Our intention is that the suffix should always communicate the purpose of the shader within the renderer, but the ":shader" suffix communicates nothing. Converting here rather than in the renderer backends (see todo in ArnoldRenderer) means we keep the dataflow in Gaffer meaningful, and allow users to use ShaderTweaks with what were once non-surface shader assignments.