ViewportGadget::SelectionScope : Don't leak GL attribute state
This is an entirely speculative fix for a text corruption bug we see very intermittently in our GLWidget overlays. My hunch is that we've been leaking some attribute state that has then caused problems with Qt's own GL drawing. In 6f5e3d325b60d1f35319533f23c76747a0cf7a4f we fixed the equivalent problem for the drawing we do to the screen, but ViewportGadget::SelectionScope is also used for "drawing" when hit-testing, and until now has not done the same.