BoxIn/Dot/Switch tests : Fix metadata serialisation bugs
The problem here was that we were removing the Serialisable flag as a means of saying "don't serialise this internal connection". But that is far too blunt of an instrument, because it also disables the serialisation of everything else about the plug, including the metadata that we're relying on to put nodules in the right place.

I've resisted the temptation to build out any public API to provide full control over this, because I'm not sure what it should look like or even if we truly need it. I don't want to commit to something too early and then regret it. Instead I've just hardcoded the behaviour we need, and added some lengthy comments about how we might go about improving the situation in the future.