BoxIO : Fix loading bug
The problem is demonstrated by the new `BoxInTest.testSerialisationWithReferenceSibling()` test case. We were attempting to use `scriptExecutedSignal()` as a way of knowing when loading was complete, and then creating our promoted plug if it was missing. This was to meet user expectations when copying out of one box and pasting into another. The problem is that `scriptExecutedSignal()` is not a reliable way of knowing when loading is complete - it is also emitted after the intermediate call to `Reference::load()`, which is before connections have been made, which causes us to wrongly conclude that we have no promoted plug.

We continue to rely on the `scriptExecutedSignal()` to add the passThrough plug when loading old scripts, because fortunately we can still reliably determine if the plug is missing, even if the signal is emitted earlier than expected. Really though, this mechanism is shaky, and as this is our own use of it anywhere in Gaffer, we may be better off finding an alternative approach and removing the signal entirely.